Moved conditional to another file
