+++ src/JWT.ts
@@ -4,6 +4,7 @@ import SignerAlg from './SignerAlgorithm'
 import { decodeBase64url, EcdsaSignature, encodeBase64url } from './util'
 import VerifierAlgorithm from './VerifierAlgorithm'
 import { JWT_ERROR } from './Errors'
+import { verifyConditionalProof } from './JWTConditional'
 
 export type Signer = (data: string | Uint8Array) => Promise<EcdsaSignature | string>
 export type SignerAlgorithm = (payload: string, signer: Signer) => Promise<string>
@@ -543,69 +544,6 @@ export async function verifyJWT(
   throw new Error(
     `${JWT_ERROR.INVALID_SIGNATURE}: JWT not valid. issuer DID document does not contain a verificationMethod that matches the signature.`
   )
-}
-
-// TODO return VerificationMethod???
-export async function verifyConditionalProof(jwt: string, signer: VerificationMethod): Promise<boolean> {
-  // validate that nested signatures are valid so that we know that each level is indirectly signing the VC
-  let decoded = decodeJWT(jwt, false)
-
-  const threshold = signer.threshold as number
-  let jwtNestedLevelCount = 1
-  let signaturesThresholdCount = 0
-  const signers: string[] = [] // string of DID URLs to the verification method or submethod
-
-  let recurse = true
-  do {
-    console.log(`verifyConitionalProof(): checking JWT at level ${jwtNestedLevelCount}`)
-    const { header, data, payload, signature } = decoded
-    let validSignatureFound = false
-    if (signer.conditionWeightedThreshold) {
-      for (const condition of signer.conditionWeightedThreshold) {
-        // TODO this should call verifyJWT() instead recursively
-        let foundSigner: VerificationMethod | undefined
-        try {
-          console.log(`testing to see if ${condition.condition.id} matches`)
-          foundSigner = await verifyJWSDecoded({ header, data, signature } as JWSDecoded, condition.condition)
-        } catch (e) {
-          if (!((e as Error).message.startsWith('invalid_signature:'))) throw e
-        }
-
-        if (foundSigner && !signers.includes(foundSigner.id)) {
-          console.log(`verifyConditionalProof(): signature valid and is unique for ${foundSigner.id}`)
-          signers.push(foundSigner.id)
-          signaturesThresholdCount += condition.weight
-          validSignatureFound = true
-
-          console.log(
-            `verifyConditionalProof(): signaturesThresholdCount ${signaturesThresholdCount} >= threshold ${threshold}`
-            )
-          if (signaturesThresholdCount >= threshold) {
-            console.log(`verifyConditionalProof(): condition valid: ${signer.id}`)
-            // TODO might need to NOT exit loop when we have recursive logic, so we can check subloops
-            return true
-          }
-        }
-      }
-    }
-
-    if (!validSignatureFound) {
-      throw new Error(
-        `${JWT_ERROR.RESOLVER_ERROR}: Invalid signature at nested level ${jwtNestedLevelCount} with signer ${signature}`
-      )
-    }
-
-    jwtNestedLevelCount++
-    if (decoded.header.cty === 'JWT') {
-      console.log(`verifyConitionalProof(): must go another level deeper to level ${jwtNestedLevelCount}`)
-      decoded = decodeJWT(payload.jwt, false)
-    } else {
-      console.log(`verifyConitionalProof(): bottom jwt = ${JSON.stringify(decoded.payload, null, 2)}`)
-      recurse = false
-    }
-  } while (recurse)
-
-  return true
 }
 
 /**
